Move Regexp check from miscellaneous to regexp category. #1170
