+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/messages.properties
@@ -5,10 +5,6 @@ todo.match=Comment matches to-do format ''{0}''.
 
 upperEll=Should use uppercase ''L''.
 
-illegal.regexp=Line matches the illegal pattern ''{0}''.
-required.regexp=Required pattern ''{0}'' missing in file.
-duplicate.regexp=Found duplicate pattern ''{0}''.
-
 translation.missingKey=Key ''{0}'' missing.
 
 missing.switch.default=switch without \"default\" clause.
