+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/messages_de.properties
@@ -5,10 +5,6 @@ todo.match=Kommentar entspricht to-do-Format ''{0}''.
 
 upperEll=Zur besseren Lesbarkeit sollte ein großes ''L'' verwendet werden.
 
-illegal.regexp=Die Zeile entspricht dem verbotenen Muster ''{0}''.
-required.regexp=Keine Zeile entspricht dem Muster ''{0}''.
-duplicate.regexp=Duplikat gefunden nach dem Muster ''{0}''.
-
 translation.missingKey=Übersetzung für Schlüssel ''{0}'' fehlt.
 
 missing.switch.default=switch ohne \"default\".
