+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/messages_es.properties
@@ -4,12 +4,6 @@ todo.match=El comentario coincide con el formato to-do ''{0}''.
 
 upperEll=Debería usar ''L'' mayúscula.
 
-
-
-illegal.regexp=La línea coincide con el patrón ilegal ''{0}''.
-required.regexp=El patron requerido ''{0}'' falta en el fichero.
-
-
 translation.missingKey=La clave ''{0}'' falta.
 
 missing.switch.default=switch sin etiqueta \"default\".
