+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/messages_fi.properties
@@ -4,8 +4,5 @@ todo.match=Kommentti on to-do formaatin ''{0}'' mukainen.
 
 upperEll=Pitää olla iso ''L''.
 
-illegal.regexp=Rivi on sääntöjenvaistaisen mallin mukainen: ''{0}''.
-
-
 translation.missingKey=Käännösavain ''{0}'' puuttuu.
