+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/messages_fr.properties
@@ -4,10 +4,6 @@ todo.match=Le commentaire correspond au format TODO ''{0}''.
 
 upperEll=Utilisez un ''L'' majuscule pour une meilleure lisibilité.
 
-illegal.regexp=La ligne correspond au motif illégal ''{0}''.
-required.regexp=Le motif ''{0}'' attendu dans le fichier est manquant.
-
-
 translation.missingKey=La traduction du message ''{0}'' est manquante.
 
 missing.switch.default=Il manque le cas \"default\" dans le bloc \"switch\".
