+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/messages_ja.properties
@@ -4,9 +4,6 @@ todo.match=コメントが to-do の形式 ''{0}'' に合致しています。
 
 upperEll=大文字の ''L'' を使用すべきです。
 
-illegal.regexp=行が不正なパターン ''{0}'' に合致します。
-
-
 translation.missingKey=キー ''{0}'' がありません。
 
 missing.switch.default="default" 節の無い switch 文です。
