+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/messages_pt.properties
@@ -1,6 +1,5 @@
 descendant.token.max=Contagem de {0} para ''{2}'' descendente ''{3}'' é excede o número máximo {1}.
 descendant.token.min=Contagem de {0} para ''{2}'' descendente ''{3}'' é menos do que o número mínimo {1}.
-illegal.regexp=A linha condiz com o padrão inválido ''{0}''.
 missing.switch.default=switch sem o ramo \"default\".
 noNewlineAtEOF=File does not end with a newline.
 todo.match=O comentário condiz com o padrão de tarefa pendente ''{0}''.
