+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/messages_tr.properties
@@ -1,5 +1,3 @@
-#Generated by ResourceBundle Editor (http://eclipse-rbe.sourceforge.net)
-
 array.type.style = Dizi parantezleri geçersiz konumda bulunuyor.
 
 descendant.token.max     = ''{2}'' elemanından türemiş ''{3}'' için toplam {0} sayısı maksimum {1} değerinden daha yüksek.
@@ -7,17 +5,11 @@ descendant.token.min     = ''{2}'' elemanından türemiş ''{3}'' için toplam {
 descendant.token.sum.max = ''{2}'' altındaki toplam {0} sayısı maksimum {1} değerinden daha yüksek.
 descendant.token.sum.min = ''{2}'' altındaki toplam {0} sayısı minimum {1} değerinden daha düşük.
 
-duplicate.regexp = Tekrarlanmýþbir kalıp bulundu: ''{0}''.
-
 final.parameter = {0} parametresi ''final'' olarak tanımlanmalı.
-
-illegal.regexp = Satırda geçersiz bir kalıp var: ''{0}''.
 
 missing.switch.default = ''default'' durumu olmayan bir ''switch'' mevcut.
 
 noNewlineAtEOF = Dosyanın sonunda yeni satır karakteri yok.
-
-required.regexp = Dosyada olması gereken ''{0}'' kalıbı yok.
 
 todo.match = Açıklamalar, ''to-do'' formatı olan ''{0}'' ile çakışıyor.
