+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/regexp/messages.properties
@@ -1,2 +1,6 @@
 regexp.exceeded=Line matches the illegal pattern ''{0}''.
 regexp.minimum=File does not contain at least {0} matches for pattern ''{1}''.
+
+illegal.regexp=Line matches the illegal pattern ''{0}''.
+required.regexp=Required pattern ''{0}'' missing in file.
+duplicate.regexp=Found duplicate pattern ''{0}''.
