+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/regexp/messages_de.properties
@@ -1,2 +1,6 @@
 regexp.exceeded=Zeile trifft auf das verbotene Muster ''{0}'' zu.
 regexp.minimum=Datei enthält weniger als die erforderlichen {0} Treffer für das Muster ''{1}''.
+
+illegal.regexp=Die Zeile entspricht dem verbotenen Muster ''{0}''.
+required.regexp=Keine Zeile entspricht dem Muster ''{0}''.
+duplicate.regexp=Duplikat gefunden nach dem Muster ''{0}''.
