+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/regexp/messages_es.properties
@@ -1,0 +1,2 @@
+illegal.regexp=La línea coincide con el patrón ilegal ''{0}''.
+required.regexp=El patron requerido ''{0}'' falta en el fichero.
