+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/regexp/messages_fi.properties
@@ -1,0 +1,1 @@
+illegal.regexp=Rivi on sääntöjenvaistaisen mallin mukainen: ''{0}''.
