+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/regexp/messages_fr.properties
@@ -1,0 +1,2 @@
+illegal.regexp=La ligne correspond au motif illégal ''{0}''.
+required.regexp=Le motif ''{0}'' attendu dans le fichier est manquant.
