+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/regexp/messages_ja.properties
@@ -1,0 +1,1 @@
+illegal.regexp=行が不正なパターン ''{0}'' に合致します。
