+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/regexp/messages_pt.properties
@@ -1,0 +1,1 @@
+illegal.regexp=A linha condiz com o padrão inválido ''{0}''.
