+++ src/main/resources/com/puppycrawl/tools/checkstyle/checks/regexp/messages_tr.properties
@@ -1,4 +1,6 @@
-#Generated by ResourceBundle Editor (http://eclipse-rbe.sourceforge.net)
-
 regexp.exceeded = Satır, geçersiz ''{0}'' kalıbıyla eşleşiyor.
 regexp.minimum  = Dosya, ''{1}'' kalıbı için en az {0} eşleşme içermiyor.
+
+illegal.regexp = Satırda geçersiz bir kalıp var: ''{0}''.
+required.regexp = Dosyada olması gereken ''{0}'' kalıbı yok.
+duplicate.regexp = Tekrarlanmýþbir kalıp bulundu: ''{0}''.
